Reorg to fixes py2 tests
